Disable debug assertions on release builds
This change also enables debug assertions on dev builds. This is a band-aid for a crash that is specifically caused by a runtime debug assertion in the wgpu library Vulkan backend. This assertion was instantly crashing Refunct when the TAS loaded.

Disabling the assertion _appears_ to not affect gameplay at all, so while this option is sort of nuclear, it works on my machine. ¯\_(ツ)_/¯